week_1/api_project/requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
requests==2.28.1
kafka-python==2.0.2